requirements.txt
fastapi
uvicorn
pydantic
python-dateutil
google-generativeai
python-dotenv
